+++ MarotoPDV/ProductModel.cs
@@ -13,7 +13,7 @@ namespace MarotoPDV
         public string Provider { get; set; }
         public DateTime ExpirationDate { get; set; }
         public int Quantity { get; set; }
-        public double PurchasePrice { get; set; }
-        public double SalePrice { get; set; }
+        public decimal PurchasePrice { get; set; }
+        public decimal SalePrice { get; set; }
     }
 }
